Update seed file

+++ db/models/contact.js
@@ -2,7 +2,7 @@ module.exports = function(sequelize, DataTypes) {
   const Contact = sequelize.define('Contact', {
     id: {type: DataTypes.INTEGER, autoIncrement: true, primaryKey: true },
     contactName: DataTypes.STRING,
-    phoneNumber: DataTypes.INTEGER,
+    phoneNumber: DataTypes.DOUBLE,
   }, {
     classMethods: {
       associate: function(models) {
